[IMP] sale_subscription: make post invoice hook really post invoice
The post invoice hook wasn't triggered when the invoicing was not
done through the `recurring_invoice_call`. This PR solve the issue
by either calling the function after the cron on all the SO that
were invoiced by the CRON call or after the SO that were invoiced
another way.

Part-of: odoo/enterprise#31055