[FIX] hr_payroll: access to payroll app posible without english language
Stepts to reproduce:

To test the issue you need a database without english language installed
create a new one with only spanish, french... with the payroll app and
try to access the app.

Issue:

We will get an error and the access to the app will be impossible unless
we install english language.

Solution:

We can't access the english language through `with_context(lang='en_US')`
if we don't have the language installed. So we can use instead the data
stored in `__range` or `batch_limit_date` in order to get the proper
date.

opw-3065958

closes odoo/enterprise#34803

X-original-commit: 4e538fe7a80b297c403b9cccbec9719849bba927
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>