[FIX] hr_payroll: access to payroll app posible without english language
Stepts to reproduce:

To test the issue you need a database without english language installed
create a new one with only spanish, french... with the payroll app and
try to access the app.

Issue:

We will get an error and the access to the app will be impossible unless
we install english language.

Solution:

We can't access the english language through `with_context(lang='en_US')`
if we don't have the language installed. So we can use instead the data
stored in `__range` or `batch_limit_date` in order to get the proper
date.

opw-3065958

closes odoo/enterprise#34847

X-original-commit: 6636c760de329a1d3dcd57fe75ce61f1219dc815
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>