[IMP] sale_subscription: increment next invoice date at invoice confirmation
Before this commit, the next invoice date was updated at invoice creation.
It triggered issues when several draft invoice were created and then deleted.
The next invoice date was not updated to remove a period.

This PR prevent to create several draft invoices for the same contract and
it increment the NID at invoice confirmation.

Taskid: 3082618
Part-of: odoo/enterprise#35014