[FIX] account_reports: fix rounding of figure_type integer
This commit fixes two bugs:

Bug 1: When displaying a value of type `integer`, the value is
incorrectly rounded to 1 decimal, when in fact it should be rounded to
0 decimal.

Solution: set the rounding to 0 digits in `format_value()` when the
figure_type is integer.

Bug 2: When displaying a value of type `integer` that is greater than
0.1 but less than 0.5, the `is_zero` property of the `column_data`
data structure generated by `_build_static_line_columns()` at
https://github.com/odoo/enterprise/blob/16.0/account_reports/models/account_report.py#L1976
is incorrectly set to False, when it should be set to True.

closes odoo/enterprise#34448

Solution: pass the `figure_type` parameter to `is_zero()`.
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>