[FIX] industry_fsm_stock: set correct qty_delivered with timesheet
Current behaviour:
Due to a recent modification (#32299), field-services that should be invoiced
based on timesheet, were invoiced based on the ordered quantity.

Expected behaviour:
Invoicing a service with an invoice policy of delivered timesheet,
should be invoiced as such, not based on orderd quantity.

Steps to reproduce:
- Install Field Server, Sales, Inventory
- Create a SO with "Field Service" as product, set the quantity to 4
- Confirm the SO
- Go to the generated task and log 1:30 hours in the timesheet.
- Validate the service
- Generate the invoice
- Observe the invoiced quantity is not the correct (4 instead of 1:30)

Reason for the problem:
The referenced change was setting manually qty_delivered, regardless
of the type of the invoicing of the field service.

Fix:
Add additional condition to not set the qty_delivered on the sol,
when that service has a service_policy of "Delivered Timesheet".

Affected versions:
- 14.0
- 15.0
- saas-15.2
- saas-15.3
- 16.0
- master

opw-3074219

closes odoo/enterprise#34557

X-original-commit: 21ba5cca5107f2ac16e956d669338ecb44622d00
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>