[IMP] *: converting custom kanban views adding a doAction button
In this commit, we will convert all the customisations of the kanban view
consist in adding an always present button in the control panel that
performs an action/calls a model's method.

To do this, we will use the new display="always" parameter applicable
to buttons in the header of kanban views. This allows us to define a
button that is always visible in a list view.

closes odoo/enterprise#34453

Taskid: 3082390
Related: odoo/odoo#106732
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>