[IMP] l10n_lu_reports_annual_vat_2023: add a PDF export
Users want to be able to have a preview of the entire report before they submit it on the ECDF portal. Clicking through the tabs of the annual declaration is not very convenient and doesn't provide a good overview.
Given the structure of the l10n_lu report, exporting it in the pdf format requires some workarounds. However, this is temporary as the entire Annual tax report will be completely reworked in master.

task-3103724

closes odoo/enterprise#36397

X-original-commit: 387f88f1baf7f9c258a98e81640268bbf0d0072b
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>