[IMP] account_reports: multi-vat cross country
The aim of this commit is to make the multi-vat feature allowing to
impact several tax report with the same tax on one invoice.

Context:
A Belgian company with a warehouse in Netherland has to fill the NL tax
report when they sell goods coming from that warehouse.
At the same time, if the goods were coming from Belgium, the company is
suppose to fill the [47] tax grid.
To avoid a lot of manual work, the customer would like to be able to put
the belgian tax grid on a NL tax.

To be consistent, if it works when applying local tag to foreign tax, it
should also work when applying foreign tag on local tax.

Before the commit:
The tax tag is applied on the journal item but the Belgian tax report is
ignoring it.
Similarly, a NL tag on a belgian tax would be ignored too.

After the commit:
The Belgian tax report is taking the journal item with the Duch tax in
account.
Similarly, the Dutch tax report is also taking the journal item with the
Belgian tax in account.

Task-id: 3196547
opw-2810712

backport of 1133fdec12585de458022e4564d692d405d5d60f

closes odoo/enterprise#38718

Related: odoo/odoo#116567
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>