[IMP] industry_fsm{_sale}: hide customer field for private task
Before this commit when task is private user still able
see and set customer on that task but if task is private
then there is no use of customer on it.

This commit hide customer field when task has not project
set on it.

task-3084850

closes odoo/enterprise#35504

Related: odoo/documentation#3311
Related: odoo/odoo#109281
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>