[IMP] project: improve UX
Before this commit:
Let's consider task A depending of task B.
Task A can be planned before task B.
Example:
Task A planned from 01/01/2020 to 05/01/2020 and
Task B planned from 06/01/2020.

In this commit:
As task A is depending of task B, task B should be
done before task A and the planned_date_begin of task A
should be greater or equal to planned_date_end of task B.

A new bootstrap alert will be displayed when a task is
planned before a task on which it depends.

task-3053344

closes odoo/enterprise#33512

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>