[REF] spreadsheet_edition,*: simplify action params
*: documents_spreadsheet,spreadsheet_dashboard_edition

The action `action_open_spreadsheet` accepts many parameters and is doing
many things based on those parameters before actually opening the spreadsheet.

Among those things:
- creating a "documents.document" if needed
- reading a "spreadsheet.template", creating a "documents.document" and
  converting the data from template to normal spreadsheet
- executing pre-processing callbacks

It is a mess and difficult to understand (lots of params, lots of
subclasses).

Basically, it's doing a lot of stuff that should be performed before
opening the action. That means it's not the action's responsibility!

With this commit, we remove all the "creation" parameters:
alwaysCreate, createInFolderId, createFromTemplateId, createFromTemplateName

Instead of a param `alwaysCreate: true` and actually creating the spreadsheet
4 steps later (if you managed to find where it is, in which subclass),
we just create the spreadsheet before opening the action.

That's a first step to clean this mess.
A next step is to clean the pre-processing callbacks. Their data and
params (isNewSpreadsheet) comes from various places and it's difficult
to follow.

closes odoo/enterprise#34567

Related: odoo/odoo#106975
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>