[IMP] helpdesk_sale_timesheet: don't ignore partner_id in SLA policy
Feature description:
Applying a customer restriction on the definition of a SLA policy
should make sure that this SLA policy should be applied only on
tickets that have that specific customer on the ticket.

Implementation details:
We are modifying the extension override of the domain used to
find the SLA policies to apply to a ticket. Before that override
was using an `OR` to extend the domain, which was ignoring the
domain for the partner selection, forcing us to have different
behaviours depending on which modules were installed.
(If the `helpdesk_sale_timesheet` module is not installed,
SLAs policies are correctly applied based on ticket's `partner_id`,
but once installed, the behaviour changes, this PR fixes that)

opw-3050470

closes odoo/enterprise#34494

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>