[IMP] *: control form and kanban rendering context
This commit is the enterprise counter part of odoo/odoo#106045,
which restricts the rendering context of kanban and form views. As
a consequence, compilers needed to be adapted to use `this` to
access the component instance, and some custom views using
implementation details in their archs needed to make them available
in their rendering context, as a temporary solution before removing
those unwanted details from the archs.

Task 3085357

closes odoo/enterprise#34493

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>