[IMP] Field Service: Improve UX
Before this commit:
- Managers don't have a measure to check if the worker has started working
on an intervention at the time they were supposed to.

- Users are confused about why tasks weren't moving to stage 'to invoice'
 when marking the task as done. It was setting misleading exceptions.

- In worksheet report and portal view, delivered_qty is displayed as
quantity if the service linked to the SOL of the task is prepaid.
But in reality, if the service is prepaid,
the customer will pay for the planned number of hours,
not the effective ones. The report should thus reflect that.
In addition, some companies may not want their customers to be aware
that they were actually charged for way more hours than needed
(or the other way around: that the company screwed up and took way more time than initially planned).

After this commit:
- The date and time on which the timer is launched and stopped are tracked
in the chatter of the task.
It's a simple measure of control for managers to check if the worker has
started working on an intervention at the time they were supposed to.
It's not a cure-all, as the worker could still start the timer from their bed,
but it's better than nothing.

- 'to invoice' stage was removed from 'Field Service' project stages.

- If the service linked to the SOL of the task is prepaid,
display the ordered quantity instead of the delivered quantity.

task-2892346

closes odoo/enterprise#29178

Related: odoo/upgrade#4088
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>