[FIX] core, *: BaseModel overrides signatures
*: hr_payroll, knowledge

The various public methods of the ORM can be override in other models,
those overrides sometime don't implement the exact same signature as the
original method in the ORM. In this work we sanitize all the overrides
to ensure a better compatibility. The background objective is to make it
possible to call any public method using kwarg: `search(domain=[...])`.

* `search`, the first parameter was renamed from `args` to `domain` in
  0e9adf7 but the overrides were not updated.
* `invalidate_models` and `invalidate_recordset`, a new `flush=True`
  parameter was introduced in 9c3b9a4 but the overrides were not
  updated.

This sanitization comes with a new linter that verifies that all
overrides of BaseModel public methods share a compatible signature.

closes odoo/enterprise#34991

Related: odoo/odoo#106999
Signed-off-by: Julien Castiaux <juc@odoo.com>