[FIX] sale_subscription: avoid posting with a void template
Posting with a void template is not supported anymore. It leads to blank or
incorrectly filled emails which are not understandable for customers. It
is now clearly forbidden. This fix avoids issues with invoice_mail_template_id
is not set.

See community PR and related commit for more details.

Task-2710804 (Mail: Clean MailThread Posting API)
Prepares Task-2088884 (Mail: Use editable computed stored fields in composer)

Part-of: odoo/enterprise#31017