[FIX] dx: port back js tooling gitignore properties
The tooling has been introduce in later version of odoo.
It is a problem when somebody switches to older branches without
disabling the tooling as it would then try to commit the tooling
config files.

We just add a few lines in the gitignore to prevent this issue.

closes odoo/enterprise#34551

X-original-commit: aba3e882298060e3a6c6ef1ae1f892aeec4ced43
Related: odoo/odoo#106945
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>