[FIX] l10n_de_reports: 2 identifiers DateV Export
We should allow different identifiers for suppliers and customers.
The system parameters are not used anymore, for the length of the
account field in the export, we now use a company setting.
We should remove the system parameters.
We also changed the constrain on `l10n_de_datev_identifier`, to avoid search in a loop.

task-3053417

closes odoo/enterprise#34984

X-original-commit: a8fcdeb29a65a4f8e6073efdee7de5fca7c7ede4
Related: odoo/upgrade#4141
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>