[FIX] l10n_de_datev_reports: 2 identifiers DateV Export
We should allow different identifiers for suppliers and customers.
The system parameters are not used anymore, for the length of the
account field in the export, we now use a company setting.
We want this in stable, so it is a new module.
We also changed the constrain in `l10n_de_reports`, to avoid search in a loop.

t-3053417

closes odoo/enterprise#34959

X-original-commit: 2075d8a301d58420c7d1b2a4e731a5edd4eec43e
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>