[IMP] sale_subscription: readonly access to date fields for simple users
Before this commit, simple salesperson could update the next invoice date or
start_date and mess up with the contract. It would put the contract in weird
states difficult to understand. As there are limited cases when these fields
should be updated by hand, we decided to limit the groups who have
write access.

taskid 3082618

closes odoo/enterprise#35014

Signed-off-by: Thibault Francois <tfr@odoo.com>
Signed-off-by: Arnaud Joset <arj@odoo.com>