[FIX] web_studio: create app with pipeline and user assignment
In studio, create a new app and a new model via the Studio home menu
On the model feature selection screen, make sure "Pipeline stages" and "Use assignment"
are checked.

Go ahead with the model creation.
Enter the kanban view.

Before this commit, there was a crash, because the automatic template included "_t"
which is not exposed by the KanbanRecord.
Also, there was no chance that this string were translated whatsoever, as the xml
node in which it is contained is not picked up by the translation algorithm.

After this commit, there is no crash, and the string is translatable, and will be so on server side.

closes odoo/enterprise#33901

Related: odoo/odoo#106599
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>