[IMP] account_invoice_extract: match only close names
There are still some issues with the partner matching on the client side.
For example, when a city or a country is mentioned in the supplier name,
it can be wrongly match to another supplier from the same place.
ex: 'Ikea Brussels' can be matched with 'City of Brussels'.

To provide this from happening, and be more selective in the matches,
several things have been done.

- 2 new fields were added on account.move to store the layout_id &
  supplier name found by the OCR. (`extract_detected_layout` and
  `extract_partner_name`)

- If no partner was found based on the VAT, we try to find one based on
  the layout (and the previously found by the OCR partner name).

- When looking for a partner from an IBAN, we now verify that the name
  of the partner is also close enough.

- To estimate if the words are close enough, we use the ratio method
  from the difflib library.

task-3084906

closes odoo/enterprise#34531

Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>