[FIX] planning: fix allocated percentage calculation on shifts
Before this commit, the allocated percentage calculation of shifts
planned on many days was wrong because it uses the old allocated
percentage as a ratio to compute the new one.

This commit updates the calculation to avoid using the old allocated
percentage to compute the new one to correctly compute based on working
schedule of the resource/company and the allocated hours set on those
shifts.

closes odoo/enterprise#34568

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>