[FIX] planning: fix allocated percentage calculation on shifts
Before this commit, the allocated percentage calculation of shifts
planned on many days was wrong because it uses the old allocated
percentage as a ratio to compute the new one.

This commit updates the calculation to avoid using the old allocated
percentage to compute the new one to correctly compute based on working
schedule of the resource/company and the allocated hours set on those
shifts.

closes odoo/enterprise#34576

X-original-commit: d33a7cfb59d2f428f1112cb19f2526eb18efc74e
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>