[IMP] accounting: invoice view modifications
Following the community PR that has removed the tag payment state when the move is not posted, the tour was broken. It checked if a payment state was present in the view for a draft bill. This PR correct that by checking the state instead of the payment state

closes odoo/enterprise#34558

Task-id: 3085423
Related: odoo/odoo#106903
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>