[FIX] knowledge: fix title saving
Purpose
=======
When rendering the article tree, if the article
record is dirty, save the article data to
render the correct/up to date article data
in the tree.

Specifications
==============
This fix is a more robust solution of a bug occurring
when updating the article permission:

When renaming an article, if you do not switch
to another article or save in general before
changing its permission, and if the new permission
moves the article in the tree, its sidebar title
comes back to its previous value.

Task-3076498

closes odoo/enterprise#34450

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>