[IMP] l10n_es_reports: update tax report wizard
With odoo/odoo#107100, we are adding new taxes with new tax tags.
Those tags are now used to fill the report lines 11 and 12 of the modelo 111 report. Line 10 is also filled automatically thanks to that.
Previously, the lines were entered manually in the report.

This commit modifies the tax report to take the new tags into account, instead of having to manually input values.

task id=3085288

closes odoo/enterprise#34846

X-original-commit: 9c793688687c756e3014e8e57e001dc5f898b572
Related: odoo/odoo#107602
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Popeler Antoine (anpp) <anpp@odoo.com>