[IMP] l10n_es_reports: update tax report wizard
With odoo/odoo#107100, we are adding new taxes with new tax tags.
Those tags are now used to fill the report lines 11 and 12 of the modelo 111 report. Line 10 is also filled automatically thanks to that.
Previously, the lines were entered manually in the report.

This commit modifies the tax report to take the new tags into account, instead of having to manually input values.

task id=3085288

closes odoo/enterprise#34800

X-original-commit: 99224eb5c374f79d0eb9484d46b19bd36f7fcabe
Related: odoo/odoo#107493
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Popeler Antoine (anpp) <anpp@odoo.com>