[FIX] spreadsheet_dashboard_edition: allow to import spreadsheet data
Usecase
=======

Fiduciaries have a "master/template" database. They copy it for each
of their customers. When they update/improve one dashboard in this master
database, they want to export it and import it to their customer's databases.

Fix
===

Currently, if you have a spreadsheet json file and you want to import it
to replace a dashboard, you have to be very carefull about what you are
doing.

You must ensure the `spreadsheet_snapshot` is also updated (or reset) and
you have to make sure there's no pending revisions. Otherwise they will
be applied on data which is unrelated.

With this commit, you can update the `data` field with a new json file
and the snapshot and the revisions will be automatically deleted, as if
the spreadsheet was brand new.

The field `spreadsheet_snapshot` is removed from the list view since
it's no longer required to update/reset it manually. And it might confuse
the user, not knowing what it is.

Task 3089052

closes odoo/enterprise#34597

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>