[IMP] mrp_workorder_plm: Add reason for step suggestion in PLM
When PLM is not installed, whenever a step instruction is made, an
activity is generated in the corresponding BoM notifying the change and
why it was made.

Though with PLM there is no necessity of an activity anymore (since an
ECO is created/updated), the reason for this change is lost in the
process, hence posting it in the step chatter.

Part-of: odoo/enterprise#35052