[FIX] web_studio: better error handling and feedback
In a few use cases, the errors triggered during studio's initialization
or usage were not handled correctly:
- at initialization, when an error made the "openStudio" action fail, the production context
was polluted with the studio key.
- at initialization, errors from the viewEditorManager's renderers were swallowed up, preventing
any sort of meaningful feedback.
- during edition, errors due to user's action on a kanban view did not automatically
undo that last action.

After this commit, all those use cases are handled in a way that makes sense and allow the user
to have the right feedback.

closes odoo/enterprise#34599

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>