[IMP] account{l10n_{lt,lu,no}}_saft: show warnings before export
There are a few issues with the SAF-T report, particularly with
the errors raised when generating it.

For now, the user has to click the SAF-T button to be aware of
potential issues/missing info.

The goal is tho display those warnings as banners when the report
is loaded, instead of waiting for the user to export it.

task-2794931

closes odoo/enterprise#37550

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Co-authored-by: John Laterre (jol) <jol@odoo.com>