[FIX] knowledge: remove extra linebreaks from templates
With Knowledge it is possible to create templates that can be utilized
in a number of other modules that features the chatter or an HTML field.

Before this commit the inserted templates were comprised of extra linebreaks
before and after its content.
Now when the template is inserted, only the linebreaks after it are added.

task-3086058

closes odoo/enterprise#35243

X-original-commit: 8a52ed14d4ad2431c9d0b8dc5285b1418bf371df
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>