[FIX] knowledge: remove extra linebreaks from templates
With Knowledge it is possible to create templates that can be utilized
in a number of other modules that features the chatter or an HTML field.

Before this commit the inserted templates were comprised of extra linebreaks
before and after its content.
Now when the template is inserted, only the linebreaks after it are added.

task-3086058

closes odoo/enterprise#34641

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>