[FIX] documents_fsm: avoid deleting non empty folders
In Odoo 15, when you activate the option to
"Centralize files attached to projects and tasks" and then make
some attachments to tasks, they will appear in the workspace
specified in the documents module. But then, when upgrading to
16, a post_init script tries to delete the workspace without taking
care of all linked documents.

This PR change the condition for fsm folders to
be deleted depending on more precise conditions.

closes odoo/enterprise#34739

X-original-commit: a16cc2da79217d8edeb4caae5e682651f18b99fe
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>