[IMP] hr_appraisal: add form view for hr.appraisal.report model
currently for the model no form view is defined and thus on clicking records from the gantt view, the record is opened with default form view, with create button. And when user clicks this create button exception is raised. to avoid, this adding form view for hr.appraisal.report model.

closes odoo/enterprise#34634

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>