[FIX] knowledge: access on restricted articles
Before this commit when an article's permission was recently changed to
Restricted and the new user didn't refresh the page, he could try to access
it.

This behavior led to restricted articles looking like they are selected in
the sidebar and unable to be hidden from the user even though they don't have
access to it.

Now, when trying to access articles and the Access Error is thrown, we open
the first article that the user can access with a full reload of the page.

task-3087960

closes odoo/enterprise#34640

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>