[FIX] account_asset: change demo data
The demo data of the deferred revenues with multi company was broken since it search for the first fixed of all companies. Also the account type for the depreciation id was wrong, it is now corrected.

This pr is related to this one: https://github.com/odoo/odoo/pull/106148

closes odoo/enterprise#34180

Task-id: 3075466
Related: odoo/odoo#106148
Signed-off-by: John Laterre (jol) <jol@odoo.com>