[FIX] social: disable warnings in kanban views
This commit comes with odoo/odoo#103722

The social kanban views use forbidden owl directives in their arch
(t-on). The presence of those directives triggers warnings in the
compiler. This commit creates an extension of the compiler to
remove the warnings. Note that in master, those forbidden directives
will be removed.

Part-of: odoo/enterprise#33740