[FIX] spreadsheet_dashboard: remove currency formats
Some of the dashboard files were using currency formats, which are
hardcoded with a given currency.

Because of that, some amounts are displayed with a $ sign, even if
the amount is in €, according to the record's currency.

It should use the automatic currency formatting of LIST functions
to use the record's currency.

This commits removes the hardcoded currency formats.
Note that the automatic format displays the decimal values while
the hardcoded format did not.
This is a functional change.

opw 3071160
opw 3058192

closes odoo/enterprise#34834

X-original-commit: 5e5580e09fdb7dbf2da3a5af979e1d6baa85998d
Related: odoo/odoo#107557
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>