[IMP] account_reports,account_accountant: add default values, accept text and bool
This commit adds support for several things:
- Editable expressions can now accept strings and boolean values.
- Editable expressions can have default values.

To define a default value for an expression, another expression needs to be added with a `_default_` prefix.
For example, if expression label is `balance`, its default expression will be called `_default_balance`.
Expressions prefixed with `_default_` are not evaluated when opening the report; we just skip them.
This way, we don't recompute them all the time for no reason.

The `account.report.external.value` is evaluated based on the lock_date + 1 day.
- `tax_lock_date` for the tax report
- `fiscalyear_lock_date` for all other reports

Hence, the default values are going to be evaluated when a user makes changes through the `account.change.lock.date` wizard.
If the lock date is not set, the default values are not going to be evaluated at all.

Options:
- date_from is either the previous lock date or the beginning of the fiscal year
- date_to is the lock date + 1 day

task-3098957

closes odoo/enterprise#36150

Related: odoo/odoo#110692
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>