[FIX] planning: prevent AttributeError during hr protected fields tests
Prior to this commit:
The written code to handle `default_planning_role_id` and `planning_role_ids`
changes was making `testWriteSelfEmployee` and `testWriteOtherEmployee`
crash with a `AttributeError` error.

After this commit:
As `default_planning_role_id` should only be an int or a record or `False`,
forcing it to `False` when not being on of the previous values will not change
the code behavior while still preventing the `AttributeError` error when
running the tests.

closes odoo/enterprise#34671

X-original-commit: 85fa5aee55740aeadd5a67cd3d15039e8130bb46
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>