[FIX] marketing_automation: disable warnings in hiearchy kanban
This commit comes with odoo/odoo#103722

The hierarchy one2many kanban view uses forbidden owl directives
and owl sub components (the HierarchyKanbanRecord) in its arch.
This will be cleaned in master, but in stable, we simply disable
the checks done by the base view. We also need to manually handle
the node HierarchyKanbanRecord in the compiler as the base view
automatically lowercases node names.

Part-of: odoo/enterprise#35065