[FIX] web_studio: kanban editor: disable compiler warnings
This commit comes with odoo/odoo#103722

In studio, the js_class attribute is ignored: views are rendered
with the base implementation. In 16.0, some archs use forbidden
owl directives (e.g. t-on-click). In their corresponding js_class,
we disabled the validation s.t. there's no warning and runbot can
be green (we can't remove them in 16.0 anymore, but it will be done
in master). As Studio doesn't use the js_class, we must disable the
warnings in the editor as well.

closes odoo/enterprise#33740

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>