[IMP] account_reports: Archive tax tags upon report changes
When unlinking or writing a tax report line, the related tags that are no longer relevant were not archived.
We now archive tags if they are still used in account.move.line and display a banner
in report that let the user know there might be a problem and invite him to correct tax tags.

closes odoo/enterprise#35195

Task: 3084028
X-original-commit: 1f627733bf1bcec25536bb6747dd2fb2e9dbd6ba
Related: odoo/odoo#108562
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>