[IMP] account_reports: Archive tax tags upon report changes
When unlinking or writing a tax report line, the related tags that are no longer relevant were not archived.
We now archive tags if they are still used in account.move.line and display a banner
in report that let the user know there might be a problem and invite him to correct tax tags.

closes odoo/enterprise#36432

Task: 3084028
X-original-commit: 988d24a90d1574f9dc90e8ca568e719bebb96f78
Related: odoo/odoo#111287
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Claire Bretton <clbr@odoo.com>