[FIX] web_studio: resilience if onchange fails when entering editor
Have a model which has complex business logic located in its compute fields.
Namely, a model which has a mandatory related field.
It can be found on mrp.workorder, where the rounding of qty_remaining depends on the product.

More generally, the onchange must crash.

Enter the edition of a formView or kanbanView without records, the first onchange is passed to get default values for the record. It crashes server side.

Before this commit, the error bubbled up.

After this commit, the edition is still possible as we made the model (relationModel and flavors)
resilient to this, as we don't care at this point what are the data since we don't operate on them.

closes odoo/enterprise#34697

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>