[ADD] l10n_mx_reports_closing: Mexico month 13 trial balance
In Mexico, they export their Trial Balance to the SAT in an XML format,
and this monthly.

At the end of the year, December will be updated as normal. But in their
closing process, accountants need to post their closing and correcting
entries to close the period. At this point, they need to make a final
extract of the Trial Balance, with only the "closing entries"
associated. This is what they call their "Month 13", so a fake period at
the end of the year used to export their closing entries.

We need a way for accountants to "tag" some Journal Entries as "closing
entries" and allow them to get a special version of their Trial Balance
to export it to the authorities.

In Odoo 15, this was done by the 'l10n_mx_reports_closing' module.
But with the new reports engine in Odoo 16, we did not port this
functionality straight away.

This commit reintroduces the l10n_mx_reports_closing module, rewritten
on top of the new reports engine.

Implementation Notes:
We create a report filter 'Month 13' which changes the report
generation.

Intuitively, when the filter is set, we restrict ourselves to the Month
13 entries within the current period.

When the filter is not set, we pretend that the Month 13 entries fall
between Dec 31 (of the year of their accounting date) and Jan 1. We
build the report with that assumption. (For example, if the current
period does not span a change of year, we don't include them, and if it
does, we include the Month 13 entries that would hypothetically fall
between Dec 31 and Jan 1.)

Comparisons between periods is supported when the 'Month 13' filter is
deactivated (in comparisons, Month 13 entries are considered as being
part of December, except for the last comparison period if it ends on
December 31.)

closes odoo/enterprise#34705

Task: 3089361
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>