[FIX] l10n_mx_edi_40: add option to l10n_mx_edi_usage for clarity
P01 for l10n_mx_edi_usage was used in cfdi 3.3 as 'por definir' (to
define) and that option existed at the time.

With the transition to cfdi 4.0, that option P01 disappeared, but
we kept it and changed it in the XML to S01, which is the 'Sin efectos
fiscales' (without fiscal effects) option.

However, this is confusing for the users as 'Sin efectos fiscales' is
not the same as 'Por Definir'.  So, in order to avoid confusion, we
changed the label for the old option to be only used in 3.3 and we
added an option S01 that will be chosen by default when cfdi 4.0 is
installed.

closes odoo/enterprise#36003

X-original-commit: 01092f42a989f89336ce6f4db552d200b9173cd1
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Josse Colpaert <jco@odoo.com>