[FIX] account_avatax_sale: drop unnecessary code
Now that the SO portal page is refreshed whenever an optional product
is added/removed or its quantity updated, the code to force the tax
recomputation whenever the option is updated can be dropped.

Indeed, taxes are already recomputed every time the page is loaded,
in the override of the /my/orders/ID route.

Also clean a bit the code while working on it.

closes odoo/enterprise#34837

X-original-commit: 3196daf7b321ce44b9c7ade8d129a23e3a8d42eb
Related: odoo/odoo#107569
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>