[FIX] account_avatax_sale: drop unnecessary code
Now that the SO portal page is refreshed whenever an optional product
is added/removed or its quantity updated, the code to force the tax
recomputation whenever the option is updated can be dropped.

Indeed, taxes are already recomputed every time the page is loaded,
in the override of the /my/orders/ID route.

Also clean a bit the code while working on it.

closes odoo/enterprise#34832

Related: odoo/odoo#107283
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>