[FIX] marketing_automation: inherit mailing domain
Setting a domain on a marketing campaign did not properly set the domain
on the activities created inside of it.

The domain is computed properly on creation, or when the activity domain
is set. But this is confusing in case the activity doesn't have
a domain as the user will see no domain applied until they leave
the modal.

task-3084676

closes odoo/enterprise#34828

X-original-commit: 3b5af758ce025ba282ffbbda8199c4ddee727e3a
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>