[FIX] stock_barcode: scan location after package
The bug this commit should fix is already fixed in v16.0, but as the
original commit (in v15.0) also adds a test who isn't here in v16.0,
this commit only add this test.

== Original commit's message ==

How to reproduce:
- Enable multilocation and package in the Inventory's settings;
- Create a new package;
- Create a product;
- In the Barcode App, create a new internal transfert;
- Scan the product then scan the package -> The package is assign as the
  result package on the line;
- Scan a location -> Traceback.

After scanned the location, it will try to get previous scanned line(s)
to apply on it the scanned location as the destination.
As a package has been scanned, it will check if some lines belong to
this package, but there is an error in the condition.

X-original-commit: f2432514d5ec80cc9a46bbe0fba05982c35667de
Part-of: odoo/enterprise#35027