[FIX] account_reports: unlink carryover when resetting tax closing move to draft
Before, when resetting a tax closing move to draft, we kept the previously created carryover values. When reposting the move, we never regenerated the carryover. It meant that if the amounts in the tax report changed before reposting the closing (to fix an error, for example), the carryover values were not consistent with it anymore.

Furthermore, when unlinking a closing move previously posted then re-doing the closing on the report, a brand new move got created, generating new carryover values at posting, hence duplicating the carryover amounts.

We fix that by always unlinking the related carryover values when resetting a closing move to draft, raising an error if this would impact a locked report.

OPW 3066409

closes odoo/enterprise#34806

X-original-commit: 559fbd9134db29eb2ed096212307b8a797f1fe94
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>