[FIX] project_enterprise: Hide planning_overlap bar
Before this commit:
When planning_overlap = False and leave_warning = True,
An alert bar is displayed to show the leave warning but
and additional empty space is shown.
It's because we inherit the view and we add the leave_warning
before the button that redirect to planning_overlap.
The button doesn't have an invisible condition when
planning_overlap = False.

In this commit:
The button should be invisible when planning_overlap = False

task-3092925

closes odoo/enterprise#34746

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>