[FIX] mrp_workorder: correctly set QC title
Problem: The title of the quality checks is not the same as the title
from the quality control point. It should use the title of the QCP if
a title is set. Otherwise, it will create a title based on the QCP type
and component

v13: https://github.com/odoo/enterprise/blob/96d26a119791471376de582bec7d9b9aba0962bb/mrp_workorder/models/quality.py#L137-L142
The logic is to use the title of the QCP if there is a QCP; otherwise,
create a title based on the QC type and component

v14: https://github.com/odoo/enterprise/blob/c62f25472dc4a63b3ab2f082ba7fcfbf0937aa9c/mrp_workorder/models/quality.py#L194

The code from v13 was refactored but used the wrong boolean condition:
Currently: If there is no QCP or if there is a component, create a title based on the QC type and component
Error: regardless of whether there is a QCP or not, it will always create a new title if there is a component
Expected: it should only create a title if there is no QCP
Proposed Change: Change the condition to use AND instead of OR --> it
  will short circuit by checking first whether there is a QCP or not.
  It will only create a new title if there is no QCP AND there exists
  a component

Steps to Reproduce on Runbot:
1. Install Manufacturing, quality > enable Work Orders
2. Create C1, storable, buy
3. Create P1, storable, manufacture
4. Create a BOM for P1 > set type to Manufacture and use C1 as a component
5. Create a work order operation for P1 using the BOM from Step 4
6. Create a Quality Check Point for P1
- Set the title
- set Operations to Manufacturing
- Set Work Order operation to the one from Step 5
- Set Type to Register Consumed Materials
- Set Product to Register to C1
7.  Create an MO for P1 , Confirm, Finish the Work order
8. View Finished Steps from the finished WO and the Title is not the same as the QCP title

opw-3032479

closes odoo/enterprise#35189

X-original-commit: 19201e45de2dc64b69a378831d25c102feca0e64
Signed-off-by: Tiffany Chang <tic@odoo.com>