[FIX] account_disallowed_expenses{_fleet}: fix several bugs
Currently, there are several bugs in the disallowed expenses report.

1. If a vehicle has several rates for the period covered by the report,
the values for the children lines are wrong.
That is because the values are grouped by rate,
which can lead to values being overwritten during post processing.

2. Due to the way line ids are generated,
trying to open the General Ledger from the caret options of account lines
would raise a UserError.
That is because the action expects the account to be the last element
of the line id, which is not always the case in this report.

3. The Journal Items caret options was removed in reportalypse,
but there was no obvious reason. It is reintroduced here.

4. The warning div that is displayed when there are multiple rates
in the report's period did not span across the report
because it was not located in the correct div.

5. The Unfold All options is activated as well.

6. Tests are added to cover most of theses cases.

task-3098491

X-original-commit: a12bee569f52f73a4b6ca16db3d8269665789634
Part-of: odoo/enterprise#35903